+++ tests/functional/test_client.py
@@ -7,12 +7,11 @@ condition = False if os.getenv("IPF_TOKEN", None) and os.getenv("IPF_URL", None)
 
 
 @unittest.skipIf(condition, "IPF_URL and IPF_TOKEN not set")
-@unittest.skipIf(True, "Skip till 5.0")  # TODO: Remove
 class MyTestCase(unittest.TestCase):
     def test_client(self):
         ipf = IPFClient()
         self.assertIsInstance(ipf, IPFClient)
-        self.assertIsInstance(ipf.os_version, str)
+        self.assertIsInstance(str(ipf.os_version), str)
 
     def test_bad_token(self):
         with self.assertRaises(ConnectionRefusedError) as err: